+++ test/mock_hardware.dart
@@ -110,6 +110,10 @@ class MockHardware implements GpioHardware {
     stateChanges = new List<StateChange>();
     disableAllInterrupts();
   }
+
+  @override
+  int gpioNum(int pinNum) =>
+      [17, 18, 27, 22, 23, 24, 25, 4, 2, 3, 8, 7, 10, 9, 11, 14, 15][pinNum];
 }
 
 class StateChange {